refactor(desktop): repatriate the desktop-only wa modules
src/shared/wa/commandPalette.ts and walkthroughDialog.ts each had exactly
one production consumer - the thin desktop wrappers. Merge each into its
wrapper (desktopCommandPalette.ts, desktopOnboarding.ts), deleting the
inner createCommandPalette/createWalkthroughDialog factory layer, the
CommandPaletteClassNames/WalkthroughDialogClassNames options (class names
inlined as literals), the DesktopCommandPaletteController and
DesktopFirstRunWalkthrough aliases, and the never-overridden
isShortcut/ariaLabel options.

Closes #8825
